Photo

+++ app/src/main/java/com/irepka3/mygarden/domain/interactor/FlowerbedInteractorImpl.kt
@@ -2,8 +2,11 @@ package com.irepka3.mygarden.domain.interactor
 
 import android.util.Log
 import com.irepka3.mygarden.domain.model.Flowerbed
+import com.irepka3.mygarden.domain.repository.DirRepository
 import com.irepka3.mygarden.domain.repository.FlowerbedRepository
 import com.irepka3.mygarden.util.Const.APP_TAG
+import java.lang.Exception
+import javax.inject.Inject
 
 /**
  * Реализация интерактора для работы с клумбами
@@ -11,18 +14,22 @@ import com.irepka3.mygarden.util.Const.APP_TAG
  *
  * Created by i.repkina on 01.11.2021.
  */
-class FlowerbedInteractorImpl(private val flowerbedRepository: FlowerbedRepository): FlowerbedInteractor {
+class FlowerbedInteractorImpl
+@Inject constructor(
+    private val flowerbedRepository: FlowerbedRepository,
+    private val dirRepository: DirRepository
+    ) : FlowerbedInteractor {
+
     override fun getFlowerbedAll(): List<Flowerbed> {
         return flowerbedRepository.getFlowerbedAll()
     }
 
     override fun getFlowerbed(flowerbedId: Long): Flowerbed? {
-        Log.d(TAG, "getFlowerBed() called with: id = $flowerbedId")
        return flowerbedRepository.getFlowerbed(flowerbedId)
     }
 
-    override fun insertFlowerbed(flowerbed: Flowerbed) {
-        flowerbedRepository.insertFlowerbed(flowerbed)
+    override fun insertFlowerbed(flowerbed: Flowerbed): Long {
+        return flowerbedRepository.insertFlowerbed(flowerbed)
     }
 
     override fun updateFlowerbed(flowerbed: Flowerbed) {
@@ -30,6 +37,12 @@ class FlowerbedInteractorImpl(private val flowerbedRepository: FlowerbedReposito
     }
 
     override fun deleteFlowerbed(flowerbed: Flowerbed) {
+        if (flowerbed.flowerbedId == null)
+            throw Exception("Invalid flowerbedId, flowerbedId is null")
+        val file = dirRepository.getFlowerbedDir(flowerbed.flowerbedId)
+        if (file.exists()) {
+            file.deleteRecursively()
+        }
         flowerbedRepository.deleteFlowerbed(flowerbed)
     }
 }
